Split emscripten and native stream implementations
